refactor: Rename renderer to preload
