+++ electron/src/preload/menu/preload-about.test.renderer.ts
