+++ electron/src/preload/menu/preload-about.ts
