+++ electron/src/preload/menu/preload-proxy-prompt.ts
