+++ electron/src/window/ProxyPromptWindow.ts
@@ -29,7 +29,7 @@ const appPath = path.join(app.getAppPath(), config.electronDirectory);
 
 const promptHtmlPath = fileUrl(path.join(appPath, 'html/proxy-prompt.html'));
 const proxyPromptWindowAllowList = [promptHtmlPath, fileUrl(path.join(appPath, 'css/proxy-prompt.css'))];
-const preloadPath = path.join(appPath, 'dist/renderer/menu/preload-proxy-prompt.js');
+const preloadPath = path.join(appPath, 'dist/preload/menu/preload-proxy-prompt.js');
 
 const windowSize = {
   HEIGHT: 350,
